fix: rename input-masks.ts to remove confusing import
